+++ UnrealEngineModLoader/UE4/CoreUObject_functions.cpp
@@ -9,6 +9,11 @@ namespace UE4
 	//Functions
 	//---------------------------------------------------------------------------
 
+
+	//---------------------------------------------------------------------------
+	//UObject Functions
+	//---------------------------------------------------------------------------
+
 	std::string UObject::GetName() const
 	{
 		auto Name = *reinterpret_cast<FName*>((byte*)this + GameProfile::SelectedGameProfile.defs.UObject.Name);
@@ -48,6 +53,21 @@ namespace UE4
 		return name;
 	}
 
+	int32_t UObject::GetIndex() const 
+	{ 
+		return Read<int32_t>((byte*)this + GameProfile::SelectedGameProfile.defs.UObject.Index); 
+	};
+
+	UClass* UObject::GetClass() const 
+	{ 
+		return Read<class UClass*>((byte*)this + GameProfile::SelectedGameProfile.defs.UObject.Class); 
+	};
+
+	UObject* UObject::GetOuter() const 
+	{ 
+		return Read<UObject*>((byte*)this + GameProfile::SelectedGameProfile.defs.UObject.Outer); 
+	};
+
 	bool UObject::IsA(UClass* cmp) const
 	{
 		for (auto super = GetClass(); super; super = static_cast<UClass*>(super->GetSuperField()))
@@ -75,6 +95,153 @@ namespace UE4
 		//fn->FunctionFlags = flags;
 	}
 
+	void UObject::ProcessEvent(class UFunction* function, void* parms)
+	{
+		return reinterpret_cast<void(*)(UObject*, class UFunction*, void*)>(GameProfile::SelectedGameProfile.ProcessEvent)(this, function, parms);
+	}
+
+	UObject* UObject::StaticLoadObject(class UClass* uclass, UObject* InOuter, const wchar_t* InName, const wchar_t* Filename, unsigned int LoadFlags, void* Sandbox, bool bAllowObjectReconciliation)
+	{
+		return reinterpret_cast<UObject * (__fastcall*)(class UClass*, UObject*, const wchar_t*, const wchar_t*, unsigned int, void*, bool)>(GameProfile::SelectedGameProfile.StaticLoadObject)(uclass, InOuter, InName, Filename, LoadFlags, Sandbox, bAllowObjectReconciliation);
+	}
+
+	UObject* UObject::CallFunctionByNameWithArguments(const wchar_t* Str, void* Ar, UE4::UObject* Executor, bool bForceCallWithNonExec)
+	{
+		return reinterpret_cast<UE4::UObject * (*)(UE4::UObject*, const wchar_t*, void*, UE4::UObject*, bool)>(GameProfile::SelectedGameProfile.CallFunctionByNameWithArguments)(this, Str, Ar, Executor, bForceCallWithNonExec);
+	}
+
+	//---------------------------------------------------------------------------
+	//UField Functions
+	//-
+
+	UField* UField::GetNext() const 
+	{ 
+		return Read<UField*>((byte*)this + GameProfile::SelectedGameProfile.defs.UField.Next); 
+	};
+
+	//---------------------------------------------------------------------------
+	//UClass Functions
+	//---------------------------------------------------------------------------
+
+	UClass* UClass::LoadClassFromString(const wchar_t* InName, bool bAllowObjectReconciliation)
+	{
+		return (UClass*)UObject::StaticLoadObject(UClass::StaticClass(), nullptr, InName, nullptr, 0, nullptr, bAllowObjectReconciliation);
+	}
+
+	//---------------------------------------------------------------------------
+	//UStruct Functions
+	//---------------------------------------------------------------------------
+
+	UStruct* UStruct::GetSuperField() const 
+	{ 
+		return Read<UStruct*>((byte*)this + GameProfile::SelectedGameProfile.defs.UStruct.SuperStruct); 
+	};
+
+	UField* UStruct::GetChildren() const 
+	{ 
+		return Read<UField*>((byte*)this + GameProfile::SelectedGameProfile.defs.UStruct.Children); 
+	};
+
+	int32_t UStruct::GetPropertySize() const 
+	{ 
+		return Read<int32_t>((byte*)this + GameProfile::SelectedGameProfile.defs.UStruct.PropertiesSize); 
+	};
+
+	//---------------------------------------------------------------------------
+	//UFunction Functions
+	//---------------------------------------------------------------------------
+
+	int32_t UFunction::GetFunctionFlags() const { return Read<int32_t>((byte*)this + GameProfile::SelectedGameProfile.defs.UFunction.FunctionFlags); };
+	void* UFunction::GetFunction() const { return Read<void*>((byte*)this + GameProfile::SelectedGameProfile.defs.UFunction.Func); };
+
+	//---------------------------------------------------------------------------
+	//UWorld Functions
+	//---------------------------------------------------------------------------
+
+	ULevel* UWorld::GetPersistentLevel() const
+	{ 
+		return Read<ULevel*>((byte*)this + GameProfile::SelectedGameProfile.defs.UWorld.PersistentLevel); 
+	};
+
+	AActor* UWorld::SpawnActor(UClass* uclass, const  FTransform* transform, const FActorSpawnParameters* params)
+	{
+		return reinterpret_cast<AActor * (*)(UWorld*, UClass*, const FTransform*, const FActorSpawnParameters*)>(GameProfile::SelectedGameProfile.SpawnActorFTrans)(this, uclass, transform, params);
+	}
+
+	//---------------------------------------------------------------------------
+	//ULevel Functions
+	//---------------------------------------------------------------------------
+
+	TArray<class AActor*> ULevel::GetWorldActors() const
+	{ 
+		return Read<TArray<class AActor*>>((byte*)this + GameProfile::SelectedGameProfile.defs.ULevel.WorldArray); 
+	};
+
+	//---------------------------------------------------------------------------
+	//Actor Functions
+	//---------------------------------------------------------------------------
+
+	FTransform AActor::GetTransform()
+	{
+		static auto fn = UObject::FindObject<UFunction>("Function Engine.Actor.GetTransform");
+		struct
+		{
+			FTransform ReturnValue;
+		}params;
+		UObject::ProcessEvent(fn, &params);
+		return params.ReturnValue;
+	}
+
+	FRotator AActor::GetActorRotation()
+	{
+		static auto fn = UObject::FindObject<UFunction>("Function Engine.Actor.K2_GetActorRotation");
+		struct
+		{
+			FRotator ReturnValue;
+		}params;
+		UObject::ProcessEvent(fn, &params);
+		return params.ReturnValue;
+	}
+
+	//---------------------------------------------------------------------------
+	//GameplayStatics Functions
+	//---------------------------------------------------------------------------
+
+	class AActor* UGameplayStatics::BeginDeferredActorSpawnFromClass(class UClass* ActorClass, const struct FTransform& SpawnTransform, ESpawnActorCollisionHandlingMethod CollisionHandlingOverride, class AActor* Owner)
+	{
+		static auto fn = UObject::FindObject<UFunction>("Function Engine.GameplayStatics.BeginDeferredActorSpawnFromClass");
+		struct
+		{
+			class UObject* WorldContextObject;
+			class UClass* ActorClass;
+			struct FTransform SpawnTransform;
+			ESpawnActorCollisionHandlingMethod CollisionHandlingOverride;
+			class AActor* Owner;
+			class AActor* ReturnValue;
+		}params;
+		params.WorldContextObject = UWorld::GetWorld();
+		params.ActorClass = ActorClass;
+		params.SpawnTransform = SpawnTransform;
+		params.CollisionHandlingOverride = CollisionHandlingOverride;
+		params.Owner = Owner;
+		UObject::ProcessEvent(fn, &params);
+		return params.ReturnValue;
+	};
+
+	class FString UGameplayStatics::GetCurrentLevelName(bool bRemovePrefixString)
+	{
+		static auto fn = UObject::FindObject<UFunction>("Function Engine.GameplayStatics.GetCurrentLevelName");
+		struct
+		{
+			class UObject* WorldContextObject;
+			bool bRemovePrefixString;
+			class FString ReturnValue;
+		}params;
+		params.WorldContextObject = UWorld::GetWorld();
+		params.bRemovePrefixString = bRemovePrefixString;
+		UObject::ProcessEvent(fn, &params);
+		return params.ReturnValue;
+	}
 
 }
