+++ UnrealEngineModLoader/UnrealEngineModLoader.vcxproj
@@ -25,6 +25,7 @@
     <ClCompile Include="UnrealEngineModLoader\dllmain.cpp" />
     <ClCompile Include="UnrealEngineModLoader\GameInfo\GameInfo.cpp" />
     <ClCompile Include="UnrealEngineModLoader\Memory\mem.cpp" />
+    <ClCompile Include="UnrealEngineModLoader\Utilities\Dumper.cpp" />
     <ClCompile Include="UnrealEngineModLoader\Utilities\EngineDefFinder.cpp" />
     <ClCompile Include="UnrealEngineModLoader\Utilities\Globals.cpp" />
   </ItemGroup>
